changed package and renamed class
- use exception instead of package-private method
- copyed private static method
